+++ git-url/tests/url.rs
@@ -1,5 +1,5 @@
 pub type Error = Box<dyn std::error::Error>;
 pub type Result = std::result::Result<(), Error>;
 
-mod expand_user;
+mod expand_path;
 mod parse;
